Added methods to postReducer to comment, reset comments, renamed blog in services to posts, was refactoring and designing ui a little

+++ part7/blog-frontend/src/components/PostContextMenu.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 
 import MoreHorizRoundedIcon from "@mui/icons-material/MoreHorizRounded";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchPosts, removePost } from "../reducers/postReducer";
+import { fetchPosts, removePost, resetPostComments } from "../reducers/postReducer";
 
 const ContextMenu = ({ id }) => {
 
@@ -25,10 +25,41 @@ const ContextMenu = ({ id }) => {
 
 	const handleDelete = () => {
 
-		dispatch(removePost({ id, token: seshUser.token }));
+		if (!confirm(`Are you sure deleting "${post.title}"?`)) return;
+
+		dispatch(removePost({ id, token: seshUser.token }))
+			.unwrap()
+			.then(() => {
+
+				notify.log("Post was successfully deleted!");
+			})
+			.catch(e => {
+
+				console.error(e);
+
+				if (e.data.error.includes("invalid token")) {
+
+					if (confirm("Invalid token. Would you like to start new user session?"))
+						handleLogout();
+
+				} else if (e.data.error.includes("token has expired")) {
+
+					alert("Your session seems to be expired, please log in again");
+					handleLogout();
+				}
+			});
 
 		handleClose();
 	};
+
+	const handleResetComments = () => {
+
+		if (confirm("Are you sure you want to reset all the comments under this post?"))
+
+			dispatch(resetPostComments({ id, token: seshUser.token }));
+	};
+
+	const disabled = !(seshUser && seshUser.id === post?.user.id);
 
 	return (<>
 		<Button
@@ -41,7 +72,8 @@ const ContextMenu = ({ id }) => {
 			open={!!anchor}
 			onClose={handleClose}
 		>
-			<MenuItem disabled={!(seshUser && seshUser.id === post?.user.id)} onClick={handleDelete}>Delete Post</MenuItem>
+			<MenuItem disabled={disabled} onClick={handleResetComments}>Reset Comments</MenuItem>
+			<MenuItem disabled={disabled} onClick={handleDelete}>Delete Post</MenuItem>
 		</Menu>
 	</>);
 };
